packages/lean/pyproject.toml
[project]
name = "lean"
version = "1.0.0"
requires-python = ">=3.13"

dependencies = [
  "google-genai>=1.31.0",
  "pygobject>=3.50.0",
  "secretstorage>=3.3.3",
]

[project.scripts]
lean = "lean:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
extend-select = ["I"]
